Code review improvements:
- Don't define USERMOD_ID_SKYSTRIP globally
- Use signed loop index to match len type
- Expose/read a socket timeout; ensure it’s applied on WiFiClient before begin()
- Guard against oversized payloads vs. JSON doc capacity
- Avoid wraparound trick in RestJsonClient ctor; initialize via reset for clarity
- Make RestJsonClient non-copyable/non-movable
- Preserve tail data in SkyModel mergeSeries
- Fix wind-direction interpolation (circular wrap-around bug)

+++ usermods/usermod_v2_skystrip/delta_view.cpp
@@ -85,7 +85,7 @@ void DeltaView::view(time_t now, SkyModel const &model, int16_t dbgPixelIndex) {
   const double step = (len > 1) ? (kHorizonSec / double(len - 1)) : 0.0;
   const time_t day = 24 * 3600;
 
-  for (uint16_t i = 0; i < len; ++i) {
+  for (int i = 0; i < len; ++i) {
     const time_t t = now + time_t(std::llround(step * i));
     int idx = seg.reverse ? (end - i) : (start + i);
